Remove unused dashboard code.
